`QueryKeyAttentionMask` -> `SelfAttentionMask`
Also move the mask to the `self_attention` module.
